requirements.txt
# web
fastapi
uvicorn
requests
bs4
beautifulsoup4
flask
pydantic
kafka_python
redis
python-dotenv

# db
DBUtils
pymysql

# llm api
langchain
langchain-community
langchain-core
zhipuai

# 通用工具
cryptography